Move task utilities to common folder
