+++ helpers/tasks/common/taskSelection.js
